requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0

# DSP / numérico
numpy==1.26.4
essentia==2.1b6.dev1389

# Persistencia (opcional, según tu helper/modelos)
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9

python-dotenv==1.0.1


python-jose[cryptography]==3.3.0